third_party/WebKit/LayoutTests/virtual/threaded/animations/KeyframeEffectReadOnly-composited-animation.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animating with KeyframeEffectReadOnly objects</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<div id="target">x</div>
<div id="targetRO">x</div>
<script>
"use strict";

promise_test(function(t) {
  var effect = new KeyframeEffect(target, { opacity: [0, 0.9] }, 1000);
  var anim = target.animate(null);
  anim.effect = effect;

  var effectRO = new KeyframeEffectReadOnly(
      targetRO, { opacity: [0, 0.9] }, 1000);
  var animRO = targetRO.animate(null);
  animRO.effect = effectRO;

  return Promise.all([anim.ready, animRO.ready]).then(function() {
    assert_true(internals.isCompositedAnimation(anim),
        "Opacity animation with KeyframeEffect should be composited");
    assert_true(internals.isCompositedAnimation(animRO),
        "Opacity animation with KeyframeEffectReadOnly should be"
        + " composited");
  });
}, "Using KeyframeEffect or KeyframeEffectReadOnly should not change whether an"
   + " animation is composited");

</script>
</body>
